+++ oviyam/mayam/trunk/src/in/raster/mayam/form/WindowingManagerPanel.java
@@ -1,41 +1,41 @@
 /* ***** BEGIN LICENSE BLOCK *****
-* Version: MPL 1.1/GPL 2.0/LGPL 2.1
-*
-* The contents of this file are subject to the Mozilla Public License Version
-* 1.1 (the "License"); you may not use this file except in compliance with
-* the License. You may obtain a copy of the License at
-* http://www.mozilla.org/MPL/
-*
-* Software distributed under the License is distributed on an "AS IS" basis,
-* WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
-* for the specific language governing rights and limitations under the
-* License.
-*
-*
-* The Initial Developer of the Original Code is
-* Raster Images
-* Portions created by the Initial Developer are Copyright (C) 2009-2010
-* the Initial Developer. All Rights Reserved.
-*
-* Contributor(s):
-* Babu Hussain A
-* Meer Asgar Hussain B
-* Prakash J
-* Suresh V
-*
-* Alternatively, the contents of this file may be used under the terms of
-* either the GNU General Public License Version 2 or later (the "GPL"), or
-* the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
-* in which case the provisions of the GPL or the LGPL are applicable instead
-* of those above. If you wish to allow use of your version of this file only
-* under the terms of either the GPL or the LGPL, and not to allow others to
-* use your version of this file under the terms of the MPL, indicate your
-* decision by deleting the provisions above and replace them with the notice
-* and other provisions required by the GPL or the LGPL. If you do not delete
-* the provisions above, a recipient may use your version of this file under
-* the terms of any one of the MPL, the GPL or the LGPL.
-*
-* ***** END LICENSE BLOCK ***** */
+ * Version: MPL 1.1/GPL 2.0/LGPL 2.1
+ *
+ * The contents of this file are subject to the Mozilla Public License Version
+ * 1.1 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ * http://www.mozilla.org/MPL/
+ *
+ * Software distributed under the License is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
+ * for the specific language governing rights and limitations under the
+ * License.
+ *
+ *
+ * The Initial Developer of the Original Code is
+ * Raster Images
+ * Portions created by the Initial Developer are Copyright (C) 2009-2010
+ * the Initial Developer. All Rights Reserved.
+ *
+ * Contributor(s):
+ * Babu Hussain A
+ * Meer Asgar Hussain B
+ * Prakash J
+ * Suresh V
+ *
+ * Alternatively, the contents of this file may be used under the terms of
+ * either the GNU General Public License Version 2 or later (the "GPL"), or
+ * the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
+ * in which case the provisions of the GPL or the LGPL are applicable instead
+ * of those above. If you wish to allow use of your version of this file only
+ * under the terms of either the GPL or the LGPL, and not to allow others to
+ * use your version of this file under the terms of the MPL, indicate your
+ * decision by deleting the provisions above and replace them with the notice
+ * and other provisions required by the GPL or the LGPL. If you do not delete
+ * the provisions above, a recipient may use your version of this file under
+ * the terms of any one of the MPL, the GPL or the LGPL.
+ *
+ * ***** END LICENSE BLOCK ***** */
 package in.raster.mayam.form;
 
 import in.raster.mayam.context.ApplicationContext;
@@ -48,7 +48,7 @@ import java.awt.Dimension;
 import java.awt.Font;
 import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
-
+import javax.swing.JOptionPane;
 
 /**
  *
@@ -56,27 +56,28 @@ import java.awt.event.KeyListener;
  * @version 0.5
  *
  */
-public class WindowingManagerPanel extends javax.swing.JPanel implements KeyListener{
-    
-    /** Creates new form WindowingManagerPanel */     
+public class WindowingManagerPanel extends javax.swing.JPanel implements KeyListener {
+
+    /** Creates new form WindowingManagerPanel */
     public WindowingManagerPanel() {
         initComponents();
         initializeDefault();
     }
-    public void initializeDefault()
-   {
+
+    public void initializeDefault() {
         jComboBox1.setSelectedIndex(0);
         setPresetTableModel();
-        jTable1.addKeyListener(this);      
+        jTable1.addKeyListener(this);
 
-   }
-    private void setPresetTableModel()
-    {
-        PresetTableModel presetTableModel=new PresetTableModel();
-        presetTableModel.setData(ApplicationContext.databaseRef.getPresetValueForModality((String)jComboBox1.getSelectedItem()));
-        jTable1.setModel(presetTableModel);
-     
     }
+
+    private void setPresetTableModel() {
+        PresetTableModel presetTableModel = new PresetTableModel();
+        presetTableModel.setData(ApplicationContext.databaseRef.getPresetValueForModality((String) jComboBox1.getSelectedItem()));
+        jTable1.setModel(presetTableModel);
+
+    }
+
     /** This method is called from within the constructor to
      * initialize the form.
      * WARNING: Do NOT modify this code. The content of this method is
@@ -170,22 +171,18 @@ public class WindowingManagerPanel extends javax.swing.JPanel implements KeyList
         );
     }// </editor-fold>//GEN-END:initComponents
 
-   
-
     private void addButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_addButtonActionPerformed
-        PresetModel presetModel=new PresetModel();
+        PresetModel presetModel = new PresetModel();
         presetModel.setPresetName("PRESETNAME");
         presetModel.setWindowLevel("0");
         presetModel.setWindowWidth("0");
-        ApplicationContext.databaseRef.insertPreset(presetModel,(String)jComboBox1.getSelectedItem() );
+        ApplicationContext.databaseRef.insertPreset(presetModel, (String) jComboBox1.getSelectedItem());
         setPresetTableModel();
 }//GEN-LAST:event_addButtonActionPerformed
 
     private void jComboBox1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jComboBox1ActionPerformed
-       setPresetTableModel();
+        setPresetTableModel();
     }//GEN-LAST:event_jComboBox1ActionPerformed
-    
-    
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton addButton;
     private javax.swing.JComboBox jComboBox1;
@@ -196,23 +193,21 @@ public class WindowingManagerPanel extends javax.swing.JPanel implements KeyList
     // End of variables declaration//GEN-END:variables
 
     public void keyTyped(KeyEvent e) {
-
     }
 
-    public void keyPressed(KeyEvent e) {       
-        if(e.getKeyCode()==KeyEvent.VK_DELETE)
-        {            
-            if(jTable1.getSelectedRow()!=-1)
-            {               
-               PresetModel presetModel=((PresetTableModel)jTable1.getModel()).getRow(jTable1.getSelectedRow());
-               ApplicationContext.databaseRef.deletePreset(presetModel);
-               setPresetTableModel();
+    public void keyPressed(KeyEvent e) {
+        if (e.getKeyCode() == KeyEvent.VK_DELETE) {
+            int resetDB = 0;
+            resetDB = JOptionPane.showConfirmDialog(this, "Are you sure want to delete the preset", "Confirmation Dialog", JOptionPane.YES_NO_OPTION);
+            if (resetDB == 0) {
+                if (jTable1.getSelectedRow() != -1) {
+                    PresetModel presetModel = ((PresetTableModel) jTable1.getModel()).getRow(jTable1.getSelectedRow());
+                    ApplicationContext.databaseRef.deletePreset(presetModel);
+                    setPresetTableModel();
+                }
             }
         }
     }
-
     public void keyReleased(KeyEvent e) {
-
     }
-    
 }
